RF: Dedicated module datalad.tests.testrepos
